requirements.txt
websocket-client==1.1.0
python-dotenv
requests